Make random library headers self-contained
