+++ inst/include/dust/random/gamma_table.hpp
@@ -2,6 +2,8 @@
 #ifndef DUST_RANDOM_GAMMA_TABLE_HPP
 #define DUST_RANDOM_GAMMA_TABLE_HPP
 
+#include "dust/random/cuda.hpp"
+
 namespace dust {
 namespace random {
