+++ inst/include/dust/random/generator.hpp
@@ -30,29 +30,10 @@
 
 #include "dust/random/cuda.hpp"
 #include "dust/random/utils.hpp"
+#include "dust/random/xoshiro_state.hpp"
 
 namespace dust {
 namespace random {
-
-// Generic data storage, this is common
-template <typename T, size_t N, xoshiro_mode M>
-struct xoshiro_state {
-  typedef T int_type;
-  HOSTDEVICE static constexpr size_t size() {
-    return N;
-  }
-  int_type state[N];
-  // This flag indicates that the distributions should return the
-  // deterministic expectation of the draw, and not use any random
-  // numbers
-  bool deterministic = false;
-  HOSTDEVICE int_type& operator[](size_t i) {
-    return state[i];
-  }
-};
-
-template <typename T>
-typename T::int_type next(T& state);
 
 template <typename T, size_t N, xoshiro_mode M>
 inline HOST
