+++ inst/include/dust/random/numeric.hpp
@@ -5,6 +5,9 @@
 #include <cfloat>
 #include <cstdint>
 #include <limits>
+#include <type_traits>
+
+#include "dust/random/cuda.hpp"
 
 namespace dust {
 namespace random {
