+++ inst/include/dust/random/xoshiro128.hpp
@@ -1,6 +1,9 @@
 #ifndef DUST_RANDOM_XOSHIRO128_HPP
 #define DUST_RANDOM_XOSHIRO128_HPP
 
+#include "dust/random/xoshiro_state.hpp"
+#include "dust/random/utils.hpp"
+
 // 32-bit generators, state is 4 * uint32_t
 //
 //  xoshiro128**  | https://prng.di.unimi.it/xoshiro128starstar.c
