+++ inst/include/dust/random/xoshiro512.hpp
@@ -1,6 +1,9 @@
 #ifndef DUST_RANDOM_XOSHIRO512_HPP
 #define DUST_RANDOM_XOSHIRO512_HPP
 
+#include "dust/random/xoshiro_state.hpp"
+#include "dust/random/utils.hpp"
+
 // 64-bit Generators, state is 8 * uint64_t
 //
 // xoshiro512**  | https://prng.di.unimi.it/xoshiro512starstar.c
